+++ shared/languageUtils/commonBlocks/getPathParamsFromPathParamsBlock.ts
@@ -3,5 +3,7 @@ import { DictionaryBlock } from "../..";
 export function getPathParamsFromPathParamsBlock(
     pathParamsBlock: DictionaryBlock,
 ) {
-    return pathParamsBlock.content.map(({ key }) => key);
+    return pathParamsBlock.content
+        .filter(({ disabled }) => !disabled)
+        .map(({ key }) => key);
 }